refactor: step1리뷰 반영
