+++ src/main/kotlin/study/Language.kt
@@ -10,14 +10,3 @@ data class Language(val type: Type, val level: Int) {
         fun english(level: Int) = Language(Type.ENGLISH, level)
     }
 }
-
-class LanguagesBuilder {
-    private val languages = mutableListOf<Language>()
-
-    infix fun String.level(value: Int) {
-        val type = Language.Type.values().first { it.name.lowercase() == this.lowercase() }
-        languages.add(Language(type, value))
-    }
-
-    fun build() = languages.toList()
-}
